+++ playwright/pageobjects/review.page.ts
@@ -7,15 +7,23 @@ export class ReviewPage {
   readonly stepHeading: Locator;
   readonly personalInfoHeading: Locator;
   readonly firstNameHeading: Locator;
+  readonly firstNameInput: Locator;
   readonly firstNameValue: Locator;
   readonly lastNameHeading: Locator;
+  readonly lastNameInput: Locator;
   readonly lastNameValue: Locator;
   readonly dateOfBirthHeading: Locator;
+  readonly dateOfBirthInput: Locator;
   readonly dateOfBirthValue: Locator;
   readonly emailHeading: Locator;
+  readonly emailInput: Locator;
   readonly emailValue: Locator;
   readonly receiveInfoHeading: Locator;
   readonly receiveInfoChoice: Locator;
+  readonly receiveInfoCheckbox: Locator;
+  readonly editPersonalInfoButton: Locator;
+  readonly alternateFormLink: Locator;
+  readonly locationsLink: Locator;
 
   constructor(page: Page) {
     this.page = page;
@@ -32,23 +40,57 @@ export class ReviewPage {
       level: 3,
     });
     this.firstNameHeading = page.getByText("First Name", { exact: true });
+    this.firstNameInput = page.getByRole("textbox", {
+      name: "First Name (Required)",
+      exact: true,
+    });
     this.firstNameValue = page.getByText(TEST_PATRON_INFO.firstName, {
       exact: true,
     });
     this.lastNameHeading = page.getByText("Last Name", { exact: true });
+    this.lastNameInput = page.getByRole("textbox", {
+      name: "Last Name (Required)",
+      exact: true,
+    });
     this.lastNameValue = page.getByText(TEST_PATRON_INFO.lastName, {
       exact: true,
     });
     this.dateOfBirthHeading = page.getByText("Date of Birth", { exact: true });
+    this.dateOfBirthInput = page.getByRole("textbox", {
+      name: "Date of Birth (Required)",
+      exact: true,
+    });
     this.dateOfBirthValue = page.getByText(TEST_PATRON_INFO.dateOfBirth, {
       exact: true,
     });
     this.emailHeading = page.getByText("Email Address", { exact: true });
+    this.emailInput = page.getByRole("textbox", {
+      name: "Email Address (Required)",
+      exact: true,
+    });
     this.emailValue = page.getByText(TEST_PATRON_INFO.email, { exact: true });
     this.receiveInfoHeading = page.getByText(
       "Receive information about NYPL's programs and services",
       { exact: true }
     );
     this.receiveInfoChoice = page.getByText("Yes", { exact: true });
+    this.receiveInfoCheckbox = page.getByText(
+      "Yes, I would like to receive information about NYPL's programs and services",
+      { exact: true }
+    );
+    this.editPersonalInfoButton = page
+      .getByRole("button", {
+        name: "Edit",
+        exact: true,
+      })
+      .first();
+    this.alternateFormLink = this.page.getByRole("link", {
+      name: "alternate form",
+      exact: true,
+    });
+    this.locationsLink = this.page.getByRole("link", {
+      name: "locations",
+      exact: true,
+    });
   }
 }