+++ playwright/tests/review.spec.ts
@@ -27,3 +27,41 @@ test.describe("displays elements on review page", () => {
   // test: displays create your account section headings
   // test: displays submit button
 });
+
+test.describe("edits patron information on review page", () => {
+  test("displays editable Personal Information section", async ({ page }) => {
+    const reviewPage = new ReviewPage(page);
+    await expect(reviewPage.editPersonalInfoButton).toBeVisible();
+    await reviewPage.editPersonalInfoButton.click();
+    await expect(reviewPage.firstNameInput).toBeVisible();
+    await expect(reviewPage.lastNameInput).toBeVisible();
+    await expect(reviewPage.dateOfBirthInput).toBeVisible();
+    await expect(reviewPage.emailInput).toBeVisible();
+    await expect(reviewPage.receiveInfoCheckbox).toBeVisible();
+    await expect(reviewPage.alternateFormLink).toBeVisible();
+    await expect(reviewPage.locationsLink).toBeVisible();
+  });
+
+  // does not replace personal info since there's no existing text
+  test("enters Personal Information", async ({ page }) => {
+    const reviewPage = new ReviewPage(page);
+    await reviewPage.editPersonalInfoButton.click();
+    await reviewPage.firstNameInput.fill(TEST_PATRON_INFO.firstName);
+    await reviewPage.lastNameInput.fill(TEST_PATRON_INFO.lastName);
+    await reviewPage.dateOfBirthInput.fill(TEST_PATRON_INFO.dateOfBirth);
+    await reviewPage.emailInput.fill(TEST_PATRON_INFO.email);
+    await reviewPage.receiveInfoCheckbox.click();
+
+    await expect(reviewPage.firstNameInput).toHaveValue(
+      TEST_PATRON_INFO.firstName
+    );
+    await expect(reviewPage.lastNameInput).toHaveValue(
+      TEST_PATRON_INFO.lastName
+    );
+    await expect(reviewPage.dateOfBirthInput).toHaveValue(
+      TEST_PATRON_INFO.dateOfBirth
+    );
+    await expect(reviewPage.emailInput).toHaveValue(TEST_PATRON_INFO.email);
+    await expect(reviewPage.receiveInfoCheckbox).not.toBeChecked();
+  });
+});